test numeric constraints
