+++ test/runtime/compiler/index.ts
@@ -3,6 +3,7 @@ import './array'
 import './boolean'
 import './date'
 import './enum'
+import './integer'
 import './intersect'
 import './keyof'
 import './literal'
