+++ test/runtime/schema/number.ts
@@ -4,8 +4,14 @@ import { ok, fail } from './validate'
 describe('type/schema/Number', () => {
   it('Should validate number', () => {
     const T = Type.Number()
+    ok(T, 3.14)
+  })
+
+  it('Should validate integer', () => {
+    const T = Type.Number()
     ok(T, 1)
   })
+
   it('Should not validate string', () => {
     const T = Type.Number()
     fail(T, 'hello')
@@ -29,5 +35,29 @@ describe('type/schema/Number', () => {
   it('Should not validate undefined', () => {
     const T = Type.Number()
     fail(T, undefined)
+  })
+
+  it('Should validate minimum', () => {
+    const T = Type.Number({ minimum: 10 })
+    fail(T, 9)
+    ok(T, 10)
+  })
+
+  it('Should validate maximum', () => {
+    const T = Type.Number({ maximum: 10 })
+    ok(T, 10)
+    fail(T, 11)
+  })
+
+  it('Should validate Date exclusiveMinimum', () => {
+    const T = Type.Number({ exclusiveMinimum: 10 })
+    fail(T, 10)
+    ok(T, 11)
+  })
+
+  it('Should validate Date exclusiveMaximum', () => {
+    const T = Type.Number({ exclusiveMaximum: 10 })
+    ok(T, 9)
+    fail(T, 10)
   })
 })
